+++ tests/test_skimming_configs.py
